Changed small things pointed out by findbugs.
git-svn-id: https://svn.origo.ethz.ch/yeti/trunk@41 7f4fa506-1231-44ca-96bd-ce53b71a5133

+++ yeti/YetiEngine.java
@@ -49,7 +49,7 @@ public class YetiEngine {
 		 */
 		public void testModuleForNMinutes(YetiModule mod, int minutes){
 			// we generate the end time using a long representation
-			long endTime = new Date().getTime()+minutes*60*1000;
+			long endTime = new Date().getTime()+(long)(minutes)*60*1000;
 			
 			// at each iteration we test the current time
 			while (new Date().getTime()<endTime){
